+++ shardingsphere-error/shardingsphere-common-error/src/main/java/org/apache/shardingsphere/error/exception/dialect/SQLDialectException.java
@@ -17,12 +17,12 @@
 
 package org.apache.shardingsphere.error.exception.dialect;
 
-import org.apache.shardingsphere.error.exception.ShardingSphereSQLException;
+import org.apache.shardingsphere.infra.util.exception.ShardingSphereInsideException;
 
 /**
  * SQL dialect exception.
  */
-public abstract class SQLDialectException extends ShardingSphereSQLException {
+public abstract class SQLDialectException extends ShardingSphereInsideException {
     
     private static final long serialVersionUID = -5090068160364259336L;
 }
